requirements.txt
appdirs==1.4.4
atpublic==2.0
backcall==0.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
configobj==5.0.6
cycler==0.10.0
cymem==2.0.3
dataclasses==0.6
decorator==4.4.2
dictdiffer==0.8.1
distro==1.5.0
dpath==2.0.1
dvc==1.6.0
fastai==2.0.0
fastcore==1.0.0
fastprogress==1.0.0
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gitdb==4.0.5
GitPython==3.1.7
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.19.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grandalf==0.6
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
jsonpath-ng==1.5.1
kiwisolver==1.2.0
matplotlib==3.3.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.2
nanotime==0.5.2
networkx==2.4
numpy==1.19.1
olefile==0.46
opencv-python==4.4.0.42
packaging==20.4
pandas==1.1.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.2.2
plac==1.1.3
ply==3.11
preshed==3.0.2
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pygtrie==2.3.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rich==5.2.1
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.2
scipy==1.5.2
setuptools==49.6.0.post20200814
shortuuid==1.0.1
shtab==1.3.1
six==1.15.0
smmap==3.0.4
spacy==2.3.2
srsly==1.0.2
tabulate==0.8.7
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
voluptuous==0.11.7
wasabi==0.7.1
wcwidth==0.2.5
wheel==0.35.1
zc.lockfile==2.0
zipp==3.1.0
